Refactor

+++ src/ui/font.rs
@@ -1,11 +1,13 @@
 #[derive(Clone)]
 pub struct FontData {
     pub data: Vec<u8>,
-    pub size: u32,
+    pub size: usize,
 }
 impl FontData {
-    pub fn new(data: Vec<u8>) -> FontData {
-        let size = data.len() as u32;
+    pub fn new(
+        data: Vec<u8>
+    ) -> FontData {
+        let size = data.len();
         return FontData {
             data,
             size,
